test(calling): review comments
